Reduce bandwidth use by sharing maven repo
By setting maven.repo.local to be inside the workspace, we lose all
the benefits of a shared maven cache on a static agent.  Setting
maven.repo.local on Kubernetes agents that are backed by an artifact
caching proxy is reasonable, since it keeps files inside the workspace
and relies on the maven cache to avoid repo.jenkins-ci.org downloads.